api/screening/requirements.txt
ultralytics>=8.2.0
pytesseract>=0.3.10
python-docx>=1.1.2
PyMuPDF>=1.24.7
Pillow>=10.0.0
